+++ src/ConfigEditor.tsx
@@ -49,6 +49,7 @@ export class ConfigEditor extends PureComponent<Props, State> {
   render() {
     const { options } = this.props;
     const { jsonData, secureJsonFields } = options;
+    jsonData.scalyrUrl = jsonData.scalyrUrl || 'https://app.scalyr.com';
     const secureJsonData = (options.secureJsonData || {}) as MySecureJsonData;
 
     return (
@@ -75,7 +76,7 @@ export class ConfigEditor extends PureComponent<Props, State> {
               labelWidth={8}
               inputWidth={20}
               onChange={this.onURLChange}
-              value={jsonData.scalyrUrl || 'https://app.scalyr.com'}
+              value={jsonData.scalyrUrl}
               placeholder="Scalyr server URL"
             />
           </div>